% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_aml.R
\name{aml_to_json}
\alias{aml_to_json}
\title{Convert AML to JSON}
\usage{
aml_to_json(aml, prettify = FALSE, indent = 4)
}
\arguments{
\item{aml}{a string, file, or URL in ArchieML format}

\item{prettify}{prettyify JSON?}

\item{indent}{if prettifying, what indent level? Passed to
\link[jsonlite]{prettify}}
}
\description{
Convert AML to JSON
}
