% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/archie_render.R
\name{archie_render}
\alias{archie_render}
\title{Render an .Rmd file, first extracting the ArchieML data from the plain text}
\usage{
archie_render(input, ...)
}
\arguments{
\item{input}{The .Rmd file}

\item{...}{parameters to be passed to \link{rmarkdown::render}}
}
\description{
Render an .Rmd file, first extracting the ArchieML data from the plain text
}
\examples{
\dontrun{
archie_render(system.file('test_archie_render.Rmd', package="rchie"),
							 output_dir=getwd())
}
}

