% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{archieml}
\alias{archieml}
\title{Inject text into an HTML document from an AML source}
\usage{
archieml(key = NULL, src = NULL, markdown = TRUE, inline = FALSE,
  fallback = NULL, elementId = NULL)
}
\arguments{
\item{key}{the key of the object in the ArchieML document. Nested elements
can be accessed with length > 1 keys such as \code{c("top-level", "next-level")}.}

\item{src}{the URL of the source document. If NULL, the widget will attempt
to use \code{params$aml_source}, which can be set in the YAML metadata of an R
Markdown document (see
\href{https://bookdown.org/yihui/rmarkdown/params-declare.html}{here}).}

\item{markdown}{Should the contents of the AML field be processed as
markdown? Defaults to TRUE, in which case text is processed with
\href{https://github.com/markdown-it/markdown-it}{markdown-it}.}

\item{inline}{If TRUE, markdown will be processed only inline, without any
\code{<p>} tags. Useful if you are inserting into a paragaph of text or figure.}

\item{fallback}{Text to use should the AML source or field not be found.}

\item{elementId}{If set, give the element this name. \code{params$aml_source}, so you
can set \code{aml_source:} in the \code{params:} section of the YAML header of a
parameterized report.}
}
\description{
This \href{https://www.htmlwidgets.org/}{htmlwidget} prints text from a field
in a source AML document.
}
\details{
A common source for the AML is a Google Document, which should be
accssed via its plain-text URL:
\code{https://docs.google.com/document/export?format=txt&id=DOCUMENT_ID}.  The
document will need to be publicly viewable for this to work.  Future versions
may enable authorized viewing via Google log-in. Note that due to security
limitations on cross-domain requests in most browsers, this will not work
when viewing a file locally, but it will when hosted on the web.
}
