% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_docx_text.R
\name{get_docx_text}
\alias{get_docx_text}
\title{Get the (plain) text of a word doc as a character string}
\usage{
get_docx_text(file)
}
\arguments{
\item{file}{A *.docx file}
}
\description{
Get the (plain) text of a word doc as a character string
}

