% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchie-package.R
\docType{package}
\name{rchie-package}
\alias{rchie-package}
\alias{rchie}
\title{A wrapper for the archieml-js}
\description{
\if{html}{\figure{logo.png}{options: align="right" width=120px style="padding-right:50px"}}
}
\details{
\strong{rchie} parses \href{http://archieml.org}{ArchieML} markup using the
\href{https://github.com/newsdev/archieml-js}{archieml-js}
library. The primary function is \link{from_aml}.  See the
\href{https://noamross.github.io/rchie/}{README} for details.

\tabular{ll}{ Package: \tab rchie\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2018-11-17\cr }
}
\references{
\url{http://archieml.org/}
}
\author{
Noam Ross
}
\keyword{JSON}
\keyword{V8}
\keyword{javascript}
