% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_aml.R
\name{from_aml}
\alias{from_aml}
\title{Import ArchieML data from a string, file, or URL}
\usage{
from_aml(aml, simplifyVector = FALSE,
  simplifyDataFrame = simplifyVector, simplifyMatrix = simplifyVector,
  flatten = FALSE, ...)
}
\arguments{
\item{aml}{a string, file, or URL in ArchieML format}

\item{simplifyVector, simplifyDataFrame, simplifyMatrix, flatten}{passed to
passed to \link[jsonlite]{fromJSON} to determine how JSON is parsed}

\item{...}{further arguments to be passed to \link[jsonlite]{fromJSON}, for
class-specific print methods}
}
\description{
This imports data from  New York Times' ArchieML formatas an R object.  Text
is parsed to JSON using archieml-js, then imported via jsonlite.
}
\examples{
from_aml(aml = "key: value")
from_aml("http://archieml.org/test/1.0/arrays.1.aml")
}
\references{
\url{http://archieml.org/}
}
