% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hello.R
\name{from_archie}
\alias{from_archie}
\title{Import archieML data from a string, file, or URL}
\usage{
from_archie(txt, ...)
}
\description{
This imports data from a string, file, or URL that is in the New York Times' ArchieML format
as an R object.  Text is parsed to JSON using archieml-js, then imported via jsonlite.
}
\references{
\url{http://archieml.org/}
}

