% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_archie.R
\name{from_archie}
\alias{from_archie}
\title{Import ArchieML data from a string, file, or URL}
\usage{
from_archie(aml, ...)
}
\arguments{
\item{aml}{a string, file, or URL in ArchieML format}

\item{...}{arguments to be passed to \link[jsonlite]{fromJSON} to determine
how JSON is parsed}
}
\description{
This imports data from  New York Times' ArchieML formatas an R object.  Text
is parsed to JSON using archieml-js, then imported via jsonlite.
}
\examples{
   from_archie(aml = 'key: value')
   from_archie('http://archieml.org/test/1.0/arrays.1.aml')
}
\references{
\url{http://archieml.org/}
}

