% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_archie.R
\name{from_archie}
\alias{from_archie}
\title{Import ArchieML data from a string, file, or URL}
\usage{
from_archie(aml, ...)
}
\arguments{
\item{aml}{a string, file, or URL in ArchieML format}

\item{...}{arguments to be passed to \link[jsonlite]{fromJSON} to determine
how JSON is parsed}
}
\description{
This imports data from  New York Times' ArchieML formatas an R object.  Text
is parsed to JSON using archieml-js, then imported via jsonlite.
}
\examples{
   from_archie(aml = 'key: value')
   from_archie('http://archieml.org/test/1.0/arrays.1.aml')

   \dontshow{

   library(testthat)
   library(jsonlite)
   context("archiml-js test suite")

   #wkdir <- getwd()
   #setwd("tests/testthat")
   #on.exit(setwd(wkdir))

   for(file in list.files(pattern = "\\\\.aml")) {
     imported <- from_archie(file)
     test_that(paste0(imported$test, " (", basename(file), ")"), {
       expect_identical(imported[-c(1, 2)], fromJSON(imported$result))
     })
   }

   for(file in list.files(pattern = "\\\\.aml")) {
     imported <- from_archie(file, simplifyVector = FALSE)
     test_that(paste0(imported$test, " (with fromJSON arguments) (", basename(file), ")"), {
       expect_identical(imported[-c(1, 2)], fromJSON(imported$result, simplifyVector=FALSE))
     })
   }


   test_that("import from URL works", {
     imported <- from_archie("http://archieml.org/test/1.0/all.0.aml")
     expect_identical(imported[-c(1, 2)], fromJSON(imported$result))
   })

   test_that("import from string works", {
     imported <- from_archie(aml = "key: value")
     expect_identical(imported, fromJSON("{\\"key\\":\\"value\\"}"))
   })

   }
}
\references{
\url{http://archieml.org/}
}

