% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/from_archie.R
\name{from_archie}
\alias{from_archie}
\title{Import ArchieML data from a string, file, or URL}
\usage{
from_archie(txt, markdown_extensions = TRUE)
}
\arguments{
\item{txt}{a string, file, or URL in ArchieML format}

\item{markdown_extensions}{Should text be pre-processed so as to make
markdown input compatible? Helps if text is auto-generated markdown from
pandoc, and enables markdown text in keys to be preserved.}
}
\description{
This imports data from  New York Times' ArchieML format
as an R object.  Text is parsed to JSON using archieml-js, then imported via jsonlite.
}
\examples{
from_archie(txt = 'key: value')
}
\references{
\url{http://archieml.org/}
}

