% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rchie-package.R
\docType{package}
\name{rchie-package}
\alias{rchie-package}
\title{Conway-Maxwell Poisson Distribution}
\description{
Provides routines for computing the density of the Conway-Maxwell Poisson
distribution and fitting parameters to data.
}
\details{
\tabular{ll}{ Package: \tab rchie\ cr Type: \tab Package\cr Version:
\tab 0.1\cr Date: \tab 2015-03-07\cr License: \tab MIT\cr }
}
\author{
Noam Ross
}
\references{
\url{http://archieml.org/}
}
\keyword{JSON}
\keyword{V8}
\keyword{javascript}

