% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchie-package.R
\docType{package}
\name{rchie-package}
\alias{rchie-package}
\alias{rchie}
\title{Import ArchieML Data}
\description{
Parses ArchieML data (\url{http://archieml.org}) using the archieml-js
library.  The primary function is \link{from_aml}.
}
\details{
\tabular{ll}{ Package: \tab rchie\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2018-11-17\cr }
}
\references{
\url{http://archieml.org/}
}
\author{
Noam Ross
}
\keyword{JSON}
\keyword{V8}
\keyword{javascript}
